requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
setuptools==68.2.2
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.41.2
